Erro ledc

+++ main/Gyro/gyro.cpp
@@ -110,7 +110,7 @@ void Gyro::update_yaw(float *yaw_param){
     notificationValue = ulTaskNotifyTake(pdTRUE, portMAX_DELAY);
 
     if (notificationValue > 1) {
-        ESP_LOGW(SPP_TAG, "Task Notification higher than 1, value: %d", notificationValue);
+        ESP_LOGW(SPP_TAG, "Task Notification higher than 1, value: %d", (int) notificationValue);
         MPU.resetFIFO();
         return;
     }
@@ -125,10 +125,10 @@ void Gyro::update_yaw(float *yaw_param){
 
     if (fifocount > kFIFOPacketSize * 2) {
         if (!(fifocount % kFIFOPacketSize)) {
-            ESP_LOGE(SPP_TAG, "Sample Rate too high!, not keeping up the pace!, count: %d", fifocount);
+            ESP_LOGE(SPP_TAG, "Sample Rate too high!, not keeping up the pace!, count: %d", (int) fifocount);
         }
         else {
-            ESP_LOGE(SPP_TAG, "FIFO Count misaligned! Expected: %d, Actual: %d", kFIFOPacketSize, fifocount);
+            ESP_LOGE(SPP_TAG, "FIFO Count misaligned! Expected: %d, Actual: %d", (int) kFIFOPacketSize, (int) fifocount);
         }
         MPU.resetFIFO();
         return;
@@ -165,7 +165,7 @@ void Gyro::read(float * pitch_param, float* yaw_param, float* roll_param){
     notificationValue = ulTaskNotifyTake(pdTRUE, portMAX_DELAY);
 
     if (notificationValue > 1) {
-        ESP_LOGW(SPP_TAG, "Task Notification higher than 1, value: %d", notificationValue);
+        ESP_LOGW(SPP_TAG, "Task Notification higher than 1, value: %d", (int) notificationValue);
         MPU.resetFIFO();
         return;
     }
@@ -180,10 +180,10 @@ void Gyro::read(float * pitch_param, float* yaw_param, float* roll_param){
 
     if (fifocount > kFIFOPacketSize * 2) {
         if (!(fifocount % kFIFOPacketSize)) {
-            ESP_LOGE(SPP_TAG, "Sample Rate too high!, not keeping up the pace!, count: %d", fifocount);
+            ESP_LOGE(SPP_TAG, "Sample Rate too high!, not keeping up the pace!, count: %d", (int) fifocount);
         }
         else {
-            ESP_LOGE(SPP_TAG, "FIFO Count misaligned! Expected: %d, Actual: %d", kFIFOPacketSize, fifocount);
+            ESP_LOGE(SPP_TAG, "FIFO Count misaligned! Expected: %d, Actual: %d", (int) kFIFOPacketSize, (int) fifocount);
         }
         MPU.resetFIFO();
         return;
